Remind to run "yarn install" if deps out of date
When developing a project, it is easy to forget to run `yarn install`
upon switching branches. If the branch has a different set of
dependencies in its package.json or yarn.lock file, the app may still
boot up, but you will likely run into errors at some point.

This commit adds a `yarn check --integrity` step, which is an extremely
fast check (~ 0.1s) to see if the current lock file and package.json
have been `yarn install`'ed. If there is a mismatch, the prestart script
will print a reminder and abort starting the app.